A whole bunch of fun stuff to build phrases into a lesson

+++ noodles/Model/Inline/Spices.swift
@@ -10,8 +10,8 @@ import Foundation
 
 /// Setting required to cook a `Bowl` of `Noodle`s
 struct Spices : Codable {
-	var musicVolume: Double = 0.5
-	var voiceVolume: Double = 0.8
+	var musicVolume: Float = 0.5
+	var voiceVolume: Float = 0.8
 	var delayBetween: Double = 3
 	var delayWithin: Double = 1
 	var randomize: Bool = true
